// client/src/utils/spain-locations.generated.ts
export const ALAVA_CITIES: CityStructure[] = [
  { city: "Área de Añana", districts: [] },
  { city: "Área de Ayala", districts: [] },
  { city: "Área de Campezo- Montaña Alavesa", districts: [] },
  { city: "Área de Salvatierra", districts: [] },
  { city: "Área de Zuya", districts: [] },
  { city: "Laguardia-Rioja Alavesa", districts: [] },
  { city: "Vitoria-Gasteiz", districts: [
    { district: "Adurtza", neighborhoods: [] },
    { district: "Aranzabela - Aranbizkarra", neighborhoods: [] },
    { district: "Aretxabaleta", neighborhoods: [] },
    { district: "Armentia - Ciudad Jardín", neighborhoods: [] },
    { district: "Betoño - Abetxuko", neighborhoods: [] },
    { district: "Casco Viejo", neighborhoods: [] },
    { district: "Centro", neighborhoods: [] },
    { district: "Coronación", neighborhoods: [] },
    { district: "El Pilar", neighborhoods: [] },
    { district: "Judimendi - Sta.Lucía", neighborhoods: [] },
    { district: "Lakua", neighborhoods: [] },
    { district: "Lakua - Arriaga", neighborhoods: [] },
    { district: "Lovaina - Aranzabal", neighborhoods: [] },
    { district: "Salburua", neighborhoods: [] },
    { district: "San Cristóbal", neighborhoods: [] },
    { district: "San Martín", neighborhoods: [] },
    { district: "Sansomendi - Ali", neighborhoods: [] },
    { district: "Santiago - El Anglo", neighborhoods: [] },
    { district: "Txagorritxu", neighborhoods: [] },
    { district: "Zabalgana - Ariznabarra", neighborhoods: [] },
    { district: "Zaramaga", neighborhoods: [] },
    { district: "Zona rural este", neighborhoods: [] },
    { district: "Zona rural noroeste", neighborhoods: [] },
    { district: "Zona rural suroeste", neighborhoods: [] }
  ] }
];

// Albacete - 6 cities
export const ALBACETE_CITIES: CityStructure[] = [
  { city: "Área de la Manchuela", districts: [] },
  { city: "Campo de Hellín", districts: [] },
  { city: "Centro", districts: [
    { district: "Albacete", neighborhoods: ["Barrios periféricos - Pedanías", "Carretas - Huerta de Marzo", "Cañicas - Canal de María Cristina", "El Pilar", "Feria", "Franciscanos", "Fátima", "Hermanos Falcó - Sepulcro Bolera", "Hospital", "Industria", "La Milagrosa - La Estrella", "La Pajarita", "Los Llanos", "Medicina", "Parque sur", "Pedro Lamata - San Pedro Mortero", "San Antonio Abad - Polígono San Antón", "San Pablo", "Santa Teresa", "Villacerrada - Centro"] },
    { district: "Área de Barrax", neighborhoods: [] },
    { district: "Área de Chinchilla de Monte-Aragón", neighborhoods: [] }
  ] },
  { city: "Corredor de Almansa", districts: [] },
  { city: "La Mancha", districts: [] },
  { city: "Sierra de Alcaraz-Segura", districts: [] }
];
// ...
export const ALMERIA_CITIES: CityStructure[] = [
  { city: "Alpujarras", districts: [
    { district: "Área de Alhama de Almería", neighborhoods: [] },
    { district: "Área de Canjáyar", neighborhoods: [] },
    { district: "Área de Laujar de Andarax", neighborhoods: [] }
  ] },
  { city: "Área de Almería", districts: [
    { district: "Almería", neighborhoods: ["Altamira - Oliveros - Barrio Alto", "Barrio San Luis - Los Molinos", "Castell del Rey", "Centro", "Centro-Rambla", "Ciudad Jardín - Tagarete", "El Alquián - San Miguel", "El Toyo", "Esperanza - Quemadero", "La Cañada - Costacabana - Loma Cabrera", "La Chanca - Pescadería", "Los Ángeles - Cruz de Caravaca", "Nueva Andalucía - Regiones", "Piedras Redondas", "Plaza de Toros", "Retamar", "San Cristóbal", "Vega de Acá - Nueva Almería - Cortijo Grande", "Villablanca - Torrecárdenas", "Zapillo"] },
    { district: "Área de Benahadux", neighborhoods: [] },
    { district: "Área de Huercal de Almería", neighborhoods: [] }
  ] },
  { city: "Bajo Almanzora", districts: [
    { district: "Área de Pulpí", neighborhoods: ["Pulpi", "San Juan de los Terreros"] },
    { district: "Cuevas del Almanzora", neighborhoods: ["Burjulu - La Mulería", "Cuevas del Almanzora", "Desert Spring Golf", "El Calón - Pozo Esparto", "Guazamara - Los Lobos", "Las Cunas - La Algarrobina", "Palomares", "Villaricos"] },
    { district: "Huércal-Overa", neighborhoods: [] }
  ] },
  { city: "Campo de Tabernas", districts: [] },
  { city: "Filabres", districts: [] },
  { city: "Levante", districts: [
    { district: "Área de Turre - Bédar", neighborhoods: ["Antas", "Bédar", "Los Gallardos", "Turre"] },
    { district: "Garrucha", neighborhoods: [] },
    { district: "Mojácar", neighborhoods: ["Mojácar Pueblo", "Pueblo Indalo-Ventanicas-El Cantal", "Zona Urb. Marina de la Torre", "Zona Vista de los Ángeles", "Zona el Palmeral"] },
    { district: "Vera", neighborhoods: ["Cala Marqués", "El Playazo", "Las Marinas", "Las Salinas", "Lomas del Mar", "Puerto Rey", "Puerto Vera", "Urb.Golf Valle Este", "Vera Playa Naturista", "Vera Pueblo"] }
  ] },
  { city: "Los Vélez", districts: [] },
  { city: "Parque Natural Cabo de Gata - Níjar", districts: [
    { district: "Cabo de Gata", neighborhoods: [] },
    { district: "Carboneras", neighborhoods: [] },
    { district: "Nijar", neighborhoods: [] }
  ] },
  { city: "Poniente", districts: [
    { district: "Adra", neighborhoods: [] },
    { district: "Berja-Balanegra", neighborhoods: ["Balanegra", "Berja"] },
    { district: "Dalias", neighborhoods: [] },
    { district: "El Ejido", neighborhoods: ["Almerimar", "Balerma", "El Ejido", "Guardias Viejas", "Las Norias", "Matagorda", "San Agustín", "Santa María del Águila"] },
    { district: "Felix-Enix", neighborhoods: [] },
    { district: "La Mojonera", neighborhoods: [] },
    { district: "Roquetas de Mar", neighborhoods: ["Aguadulce", "El Parador", "Roquetas"] },
    { district: "Vícar", neighborhoods: ["La Envía", "La Gangosa", "Las Cabañuelas", "Puebla de Vícar", "Vícar"] }
  ] },
  { city: "Valle de Almanzora", districts: [
    { district: "Área de Albox", neighborhoods: [] },
    { district: "Área de Purchena", neighborhoods: [] },
    { district: "Área de Serón", neighborhoods: [] },
    { district: "Área de Zurgena", neighborhoods: [] }
  ] }
];

// Andorra - 7 cities
export const ANDORRA_CITIES: CityStructure[] = [
  { city: "Andorra la Vella", districts: [] },
  { city: "Canillo", districts: [] },
  { city: "Encamp", districts: [] },
  { city: "Escaldes-Engordany", districts: [] },
  { city: "La Massana", districts: [] },
  { city: "Ordino", districts: [] },
  { city: "Sant Julia de Loria", districts: [] }
];

// Asturias - 8 cities
export const ASTURIAS_CITIES: CityStructure[] = [
  { city: "Área de Avilés", districts: [
    { district: "Área de Corvera - Candamo", neighborhoods: [] },
    { district: "Área de Muros de Nalón", neighborhoods: [] },
    { district: "Avilés", neighborhoods: [] },
    { district: "Castrillon", neighborhoods: [] },
    { district: "Cudillero", neighborhoods: [] },
    { district: "Gozón", neighborhoods: [] },
    { district: "Pravia", neighborhoods: [] }
  ] },
  { city: "Área de Gijón", districts: [
    { district: "Carreño", neighborhoods: [] },
    { district: "Gijón", neighborhoods: ["Centro", "El Llano", "Este", "Oeste", "Periurbano", "Somió - Cabueñes", "Sur"] },
    { district: "Villaviciosa", neighborhoods: [] }
  ] },
  { city: "Área de Narcea", districts: [] },
  { city: "Área de Oviedo", districts: [
    { district: "Área de Grado", neighborhoods: [] },
    { district: "Área de Nava", neighborhoods: [] },
    { district: "Área de Somiedo", neighborhoods: [] },
    { district: "Llanera", neighborhoods: [] },
    { district: "Noreña", neighborhoods: [] },
    { district: "Oviedo", neighborhoods: ["Buenavista-Ería-Montecerrao", "Centro-Casco Histórico", "Ciudad Naranco-Vallobín-La Florida", "Corredoria-Huca-Teatinos", "Fozaneldi-Tenderina-Ventanielles", "La Manjoya-Parroquias Sur", "Parroquias Norte-Naranco", "San Claudio-Trubia-Las Caldas", "San Lazaro-Otero-Villafría"] },
    { district: "Siero", neighborhoods: [] }
  ] },
  { city: "Caudal", districts: [
    { district: "Aller", neighborhoods: [] },
    { district: "Lena", neighborhoods: [] },
    { district: "Mieres del Camino", neighborhoods: [] }
  ] },
  { city: "Eo-Navia", districts: [
    { district: "Área de Eo", neighborhoods: [] },
    { district: "Área de Oscos - Parque Prehistórico", neighborhoods: [] },
    { district: "Área de Tapia", neighborhoods: [] },
    { district: "Área del Valle de Navia", neighborhoods: [] },
    { district: "Luarca - Valdés", neighborhoods: [] }
  ] },
  { city: "Nalón", districts: [
    { district: "Langreo", neighborhoods: [] },
    { district: "Laviana", neighborhoods: [] },
    { district: "San Martín del Rey Aurelio", neighborhoods: [] },
    { district: "Sobrescobio-Caso", neighborhoods: [] }
  ] },
  { city: "Oriente", districts: [
    { district: "Área de Cangas de Onís", neighborhoods: [] },
    { district: "Área de Peñamellera", neighborhoods: [] },
    { district: "Área de Piloña", neighborhoods: [] },
    { district: "Cabrales", neighborhoods: [] },
    { district: "Colunga - Caravia", neighborhoods: [] },
    { district: "Llanes", neighborhoods: ["Celorio-Poó-Parres", "Cué-San Roque-Andrín", "Llanes", "Posada-Barro", "Pría-Nueva-Hontoria-Naves", "Vibaña-Ardisana-Caldueño", "Vidiago-Pendueles-Tresgrandas"] },
    { district: "Parres", neighborhoods: [] },
    { district: "Ponga - Amieva", neighborhoods: [] },
    { district: "Ribadedeva", neighborhoods: [] },
    { district: "Ribadesella", neighborhoods: [] }
  ] }
];

// Ávila - 5 cities
export const AVILA_CITIES: CityStructure[] = [
  { city: "Área de Arévalo", districts: [] },
  { city: "Área de Ávila", districts: [
    { district: "Área de San Pedro del Arroyo", neighborhoods: [] },
    { district: "Área de Villatoro", neighborhoods: [] },
    { district: "Ávila", neighborhoods: [] }
  ] },
  { city: "Burgohondo - Cebreros - El Tiemblo", districts: [
    { district: "Área de Burgohondo - El Barraco", neighborhoods: [] },
    { district: "Área de Cebreros-Hoyo de Pinares", neighborhoods: [] },
    { district: "El Tiemblo", neighborhoods: [] },
    { district: "Navalperal-Las Navas del Marqués-Peguerinos", neighborhoods: [] }
  ] },
  { city: "Valle del Tiétar", districts: [
    { district: "Área de Arenas de San Pedro", neighborhoods: [] },
    { district: "Área de Casa Vieja - Gavilanes", neighborhoods: [] },
    { district: "Área de la Adrada - Piedralaves", neighborhoods: [] },
    { district: "Área de Sotillo de la Adrada", neighborhoods: [] },
    { district: "Candeleda", neighborhoods: [] }
  ] },
  { city: "Valle del Tormes", districts: [] }
];
// ...
export const CEUTA_CITIES: CityStructure[] = [
  { city: "Hacho - San Amaro", districts: [] },
  { city: "Zona Centro", districts: [] },
  { city: "Zona Muelle - Extrarradio", districts: [] }
];
// ...
export const SPAIN_LOCATIONS: CityStructure[] = [
  ...ALAVA_CITIES,
  ...ALBACETE_CITIES,
  ...ALICANTE_CITIES,
  ...ALMERIA_CITIES,
  ...ANDORRA_CITIES,
  ...ASTURIAS_CITIES,
  ...AVILA_CITIES,
  ...BADAJOZ_CITIES,
  ...BALEARS___ILLES__CITIES,
  ...BARCELONA_CITIES,
  ...BURGOS_CITIES,
  ...CACERES_CITIES,
  ...CADIZ_CITIES,
  ...CANTABRIA_CITIES,
  ...CASTELLON_CITIES,
  ...CEUTA_CITIES,
  ...CIUDAD_REAL_CITIES,
  ...CORDOBA_CITIES,
  ...CUENCA_CITIES,
  ...GIPUZKOA_CITIES,
  ...GIRONA_CITIES,
  ...GRANADA_CITIES,
  ...GUADALAJARA_CITIES,
  ...HUELVA_CITIES,
  ...HUESCA_CITIES,
  ...JAEN_CITIES,
  ...LA_RIOJA_CITIES,
  ...LAS_PALMAS_CITIES,
  ...LEON_CITIES,
  ...LLEIDA_CITIES,
  ...LUGO_CITIES,
  ...MADRID_CITIES,
  ...MALAGA_CITIES,
  ...MELILLA_CITIES,
  ...MURCIA_CITIES,
  ...NAVARRA_CITIES,
  ...OURENSE_CITIES,
  ...PALENCIA_CITIES,
  ...PONTEVEDRA_CITIES,
  ...SALAMANCA_CITIES,
  ...SANTA_CRUZ_DE_TENERIFE_CITIES,
  ...SEGOVIA_CITIES,
  ...SEVILLA_CITIES,
  ...SORIA_CITIES,
  ...TARRAGONA_CITIES,
  ...TERUEL_CITIES,
  ...TOLEDO_CITIES,
  ...VALENCIA_CITIES,
  ...VALLADOLID_CITIES,
  ...VIZCAYA_CITIES,
  ...ZAMORA_CITIES,
  ...ZARAGOZA_CITIES
];